adding classes FieldStatistics and FieldStatisticsConfig. This assumes that we are only interested in collecting stats for individual fields. Check if this is the case
Changed the structure containing the stats info from OperatorStatisticsConfig to an array of FieldStatisticsConfig. Still missing how to configure the stats in the conf HashMap

FieldStatistics now have a function and fieldName fields that define what element to apply the stats to. Still need to configure the call to the key selector function in the OperatorStatistics process function

removed OperatorStatisticsConfig.

added taskName to identify which task the OperatorStatistics are associated to. The task name comes from the task config