[FLINK-1297] Added an accumulator called OperatorStatsAccumulator capable of tracking min, max and estimates for count distinct and heavy hitters.
The count distinct algorithms are Linear Counting and HyperLogLog, both from an imported library from clearspring.

The heavy hitters algorithms are Lossy counting (Manku et.al 2002) and one based on Count Min Sketch (Cormode 2005).

The heavy hitters algorithms are implemented in the statistics package in flink-contrib.

Refactored to flink-contrib.

The OperatorStatsAccumulatorTest creates a local accumulator and a global accumulator. Inside of the
flatmap the local accumulator is updated. The global accumulator is updated when the task closes.

Custom serialization implemented for OperatorStatistics and CountMinHeavyHitter.
Refactored CountMinHeavyHitter and fixed a bug in the merge function.
Fixed bug in merge function in OperatorStatistics.